Changed the DNode shutdown hook
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# En la rama master
# Your branch is up-to-date with 'origin/master'.
#
# Cambios para hacer commit:
#	modificado: splout-hadoop/src/main/java/com/splout/db/hadoop/SchemaSampler.java
#
# Archivos sin seguimiento:
#	CHANGES.txt~
#	README.md~
#	out-com.splout.db.hadoop.TestTablespaceGenerator/
#	out-com.splout.db.hadoop.TestTablespaceGeneratorJavaScript/
#	splout-hadoop/.in-com.splout.db.hadoop.TestTablespaceGeneratorJavaScript.crc
#	splout-hadoop/.input-com.splout.db.hadoop.TestTupleSampler_1.crc
#	splout-hadoop/.input-com.splout.db.hadoop.TestTupleSampler_2.crc
#	splout-hadoop/.input-com.splout.db.hadoop.TestTupleSampler_3.crc
#	splout-hadoop/.input-com.splout.db.hadoop.TestTupleSampler_r.crc
#	splout-hadoop/.output-com.splout.db.hadoop.TestTupleSampler_1.crc
#	splout-hadoop/.output-com.splout.db.hadoop.TestTupleSampler_2.crc
#	splout-hadoop/.output-com.splout.db.hadoop.TestTupleSampler_3.crc
#	splout-hadoop/.output-com.splout.db.hadoop.TestTupleSampler_r.crc
#	splout-hadoop/database-files/
#	splout-hadoop/ddate.json
#	splout-hadoop/derby.log
#	splout-hadoop/idadvertiser.json
#	splout-hadoop/idpublisher.json
#	splout-hadoop/metastore_db/
#	splout-hadoop/pangool-instances/
#	splout-hadoop/src/main/java/com/splout/db/benchmark/TablespaceAnalyserCMD.java
#	splout-server/dnode-staging-0/
#	splout-server/dnode-staging-1/
#	splout-server/dnode-staging-2/
#	splout-server/hz-data/
#

+++ splout-server/src/main/java/com/splout/db/qnode/QNodeHandlerContext.java
@@ -74,13 +74,13 @@ public class QNodeHandlerContext {
   private ReentrantLock thriftClientCacheLock = new ReentrantLock();
 
   private final int thriftClientPoolSize;
-  private final int dnodePoolTimeoutMillis;
+  private final long dnodePoolTimeoutMillis;
 
   public QNodeHandlerContext(SploutConfiguration config, CoordinationStructures coordinationStructures) {
     this.config = config;
     this.coordinationStructures = coordinationStructures;
     this.thriftClientPoolSize = config.getInt(QNodeProperties.DNODE_POOL_SIZE);
-    this.dnodePoolTimeoutMillis = config.getInt(QNodeProperties.QNODE_DNODE_POOL_TAKE_TIMEOUT);
+    this.dnodePoolTimeoutMillis = config.getLong(QNodeProperties.QNODE_DNODE_POOL_TAKE_TIMEOUT);
     this.replicaBalancer = new ReplicaBalancer(this);
     initMetrics();
   }
